usbmuxd image: Build an -udev variant
By default, we build libusb with `--enable-udev=no`.
On Linux environments where usbfs is not available, `libusb_init` will
fail and usbmuxd will exit with a SIGSEV.

While it's okay for usbmuxd to fail when no USB stack is available
(no iOS devices are present either), this causes issues in CI systems,
where we'd prefer for usbmuxd to run but just report an empty device
list.

As a workaround, we can compile libusb with udev support. This will
cause libusb to report an empty device list and usbmuxd to work
correctly.

Because we don't want udev support in the main container image,
this introduces a new tag, latest-udev-amd64.

The Dockerfiles for usbmuxd with and without udev support are similar,
and are built from a template using gomplate.