Add Kubernetes watcher logic
The standard Watch implementation in the .NET KubernetesClient suffers from a couple of issues:

- [The first event is ignored](https://github.com/kubernetes-client/csharp/issues/551)
- [The watcher stops watching](https://github.com/kubernetes-client/csharp/issues/533)
- There's a non-trivial API involved, while watching essentially is a `Task` which reads data off a `Stream`.

This commit introduces the `KubernetesClient` class which inherits from `Kubernetes`, and adds a `WatchPodAsync` method, which allows you to observe pod events.

It also includes a workaround for [another issue in the Kubernets client, where the 0.0.0.0 IP address cannot be used](https://github.com/kubernetes-client/csharp/pull/550).