Vagrant box: Use k3s instead of k3d
The Vagrant box was using k3d (k3s-in-docker) to provision k3s. As a result, pods run
inside containerd inside Docker. A limitation of this setup is that usb
hotplug does not work correctly: newly attached devices are not discoverable
inside the container.

This PR reworks the Vagrant setup to use "native" k3s, and forces k3s to use
Docker as the container runtime (simplifying the setup a bit).