iOS: Support registering clients via dependency injection
This commit adds the ability to register services which interact with devices via dependency injection. It introduces the concept of a "device scope".

A device scope is a scope in which the DeviceContext service is registered. This service contains the information about the device on which the scope interacts.

You can configure a service collection by calling the `AddAppleServices` method. Once the service provider has been built, you can create a device scope by calling the `CreateDeviceScopeAsync` method.

This method takes a `udid` as a parameter (which can be `null` or a specific UDID).