Deploy a Docker registry as part of Kaponata
We will store the iOS developer disk images as container images in the Docker registry.

By default, the registry will use "local-path" storage. This is
(non-redundant) cluster-local storage and is enable by default
in k3s clusters.

The registry is deployed as a StatefulSet to ensure the PVC
can be reused after redeploying the same helm chart.