Introducing the DeviceServiceProvider
The `DeviceServiceProvider` class is a simple class which wraps around the `IServiceProvider` and exposes methods to interact with lockdown services running on iOS devices.

The `CreateDeviceScopeAsync` and `StartServiceAsync` methods replace the extension methods on the `IServiceProvider` interface, making these methods easier to mock.

There are no function changes in this PR.