Use a dedicated DeviceServiceScope
Interacting with lockdown services on iOS devices requires you to create a `IServiceScope`. This `IServiceScope` must have a `DeviceContext` service which contains a at least the `MuxerDevice` object, a `MuxerClient` service and a factory which can create your service.

We previously implemented this using extension methods on the `IServiceScope` interface, but these are not easy to mock.

This PR rewrites the functionality to use a dedicated `DeviceServiceScope` object, from which you can source the required services.