List & Watch operations: Support the resourceVersionMatch parameter
Various operations (such as the list and watch operations) take a `resourceVersion` argument, which allows you to specify which version of a resource you want to retrieve.

It is possible for exact resource versions of objects the be unavailable, when these objects are not updated frequently.

In that case, we can pass `resourceVersionMatch =NotOlderThan`, allowing the cluster to return newer versions of that object.