UsbmuxdSidecar: Use the UDID of the device as a unique identifier
The sidecar currently selects iOS devices based on the OS, the UDID and the "owner" (= the name of the usbmuxd pod to which the device is attached).

That doesn't work well for devices which were previously attached to a pod with a different name. This can happen when the pod is re-created, or when the device is physically moved from one node to another.

Only use the OS + the UDID of the device to uniquely identifier a device.

This commit also adds more logging.